Macathon/Macathon/requirements.txt
# Core runtime for Toronto tracker backend
python-dotenv
pydantic
fastapi
uvicorn[standard]
google-genai
